+++ packages/stream_chat_flutter/lib/src/stream_chat.dart
@@ -126,11 +126,25 @@ class StreamChatState extends State<StreamChat> {
     return defaultTheme.merge(themeData);
   }
 
+  // coverage:ignore-start
+
   /// The current user
-  User? get user => widget.client.state.user;
+  @Deprecated('Use `.currentUser` instead, Will be removed in future releases')
+  User? get user => widget.client.state.currentUser;
 
   /// The current user as a stream
-  Stream<User?> get userStream => widget.client.state.userStream;
+  @Deprecated(
+    'Use `.currentUserStream` instead, Will be removed in future releases',
+  )
+  Stream<User?> get userStream => widget.client.state.currentUserStream;
+
+  // coverage:ignore-end
+
+  /// The current user
+  User? get currentUser => widget.client.state.currentUser;
+
+  /// The current user as a stream
+  Stream<User?> get currentUserStream => widget.client.state.currentUserStream;
 
   @override
   void initState() {