+++ packages/stream_chat_flutter/test/src/date_divider_test.dart
@@ -13,7 +13,7 @@ void main() {
       final clientState = MockClientState();
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(OwnUser(id: 'user-id'));
+      when(() => clientState.currentUser).thenReturn(OwnUser(id: 'user-id'));
 
       await tester.pumpWidget(MaterialApp(
         home: StreamChat(