+++ packages/stream_chat_flutter_core/lib/src/message_list_core.dart
@@ -116,7 +116,7 @@ class MessageListCoreState extends State<MessageListCore> {
 
   bool get _isThreadConversation => widget.parentMessage != null;
 
-  OwnUser? get _currentUser => _streamChannel!.channel.client.state.user;
+  OwnUser? get _currentUser => _streamChannel!.channel.client.state.currentUser;
 
   var _messages = <Message>[];