+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -6,6 +6,11 @@
 - Allow the various ListView widgets to be themed via ThemeData classes
 - Added `bottomRowBuilder` and `deletedBottomRowBuilder` that build a widget below a `MessageWidget`
 
+🔄 Changed
+
+- `StreamChat.of(context).user` is now deprecated in favor of `StreamChat.of(context).currentUser`.
+- `StreamChat.of(context).userStream` is now deprecated in favor of `StreamChat.of(context).currentUserStream`.
+
 🐞 Fixed
 
 - Fix floating date divider not having a fixed size